[HOT FIX] Gravity effector had a bad unsigned integer subtraction that was periodically nuking the perturbed gravity computation.
Former-commit-id: d3180b9594cedaed70dd3792be44047e314da9c5
Former-commit-id: ca88ecf5708ca7de0e1684ee146505658ecc307d

+++ SimCode/dynamics/_GeneralModuleFiles/gravityEffector.cpp
@@ -299,7 +299,7 @@ Eigen::Vector3d GravBodyData::computeGravityInertial(Eigen::Vector3d r_I,
     
     if(this->spherHarm.harmReady() && this->useSphericalHarmParams)
     {
-        double dt = (simTimeNanos - this->localHeader.WriteClockNanos)*NANO2SEC;
+        double dt = ((int64_t) simTimeNanos - (int64_t) this->localHeader.WriteClockNanos)*NANO2SEC;
         Eigen::Matrix3d dcm_PfixN = Eigen::Map<Eigen::Matrix3d>
             (&(this->localPlanet.J20002Pfix[0][0]), 3, 3);
         Eigen::Matrix3d dcm_PfixN_dot = Eigen::Map<Eigen::Matrix3d>
